adjust reported metrics to use new load-factor / load stats (#30)

+++ src/main/java/org/commonjava/cdi/util/weft/WeftExecutorService.java
@@ -11,7 +11,7 @@ public interface WeftExecutorService
 
     double getLoadFactor();
 
-    int getCurrentLoad();
+    long getCurrentLoad();
 
     Integer getThreadCount();
